fix(audit): /code-audit pass 4 — Pass-3 regressions + targeted polish
LOW (pre-existing or my own Pass-3 introductions):
- dbusmenumodel.cpp: 829 → 648 lines. Pass-3 additions pushed past the
  CLAUDE.md 800-line cap. Extract `unpackLayoutVariant`,
  `labelFromProps`, `shortcutFromProps`, `iconToDataUrl`,
  `iconFromProps`, `dbusmenuTimestamp` into a new
  `src/dbusmenuhelpers.{h,cpp}` translation unit and register it in
  CMakeLists.txt.
- dbusmenumodel.cpp: DRY. Pass-3 introduced a second function-local
  `static const QSet<QString> kIconProps{...}` inside the sameShape
  refresh branch alongside the original in `onPropertiesUpdated`.
  Replace both with a shared anonymous-namespace `iconPropKeys()`.
- dbusmenumodel.cpp setRootId / refresh: skip the trailing
  `d->refresh()` when `flushPendingRebuild()` actually ran a rebuild.
  Pre-fix, the flush already issued GetLayout for the new rootId and
  the explicit `d->refresh()` cancelled the just-issued watcher and
  fired a duplicate. flushPendingRebuild now returns bool (marked
  [[nodiscard]]); the no-op-on-flush callers wrap in `(void)`.
- mprisplayer.cpp Rate: clamp to [-64.0, 64.0]. Pass-1 added the
  isfinite guard but left Rate unbounded; a hostile player publishing
  Rate ~= 1e15 would survive isfinite, then
  `rate * kPositionPollMs * 1000` in tickPosition overflows qint64 on
  the static_cast (implementation-defined per [conv.fpint]). 64x
  covers every realistic fast-forward / reverse-playback case
  mainline players use.
- examples/phosphor-shell/TopPanel.qml battery row: drop the dead
  `&& root.batteryPercent.length > 0` guard. shell.qml's binding
  renders "" only while displayDevice is null and "0" + "%" once it
  appears (UPowerDevice initialises percentage = 0.0), so the length
  check is never the deciding gate under the current binding
  contract; batteryVisible alone is correct. Pass-3 rewrote the
  comment to describe a race the binding actually doesn't produce.
- examples/phosphor-shell/MprisWidget.qml MprisPlayerState comment:
  tighten the wording attribution. The savings credited to sampling
  are the `Math.min/Math.max` in MprisPlayerState's progress binding;
  the Canvas requestPaint cost is already gated by `prog`'s
  short-circuit and the `if (root.visible)` guard.
- examples/phosphor-shell/MprisWidget.qml progressRing.prog: document
  that the `root.visible && root.hasPlayer` guard is intentionally
  layered with MprisPlayerState's own sampling check (avoids a brief
  spurious paint during the visible->hidden transition while
  playerState.sampling has not yet propagated).
- Em-dash sweep across phosphor-service-icontheme src/headers/tests
  (Pass-1/2 swept CMakeLists + READMEs; the headers/src/tests were
  missed). Pass-3 did the equivalent sweep for phosphor-service-
  upower; this finishes the partition.
- Em-dash sweep across phosphor-service-mpris src/tests (Pass-3
  commit log explicitly only swept upower; mpris was missed).

Build + 208 tests green.
